Move CA conversion logic into its own module
